Update patch set 3
Patch Set 3:

this looks good so far. A potential addition to this change is to define 1G hugepages as default. I have seen issues with 1G default hugepages in my Airship dev environment and would like to check the behavior of this version of Airship in this environment.

Patch-set: 3
Label: Verified=0