use uuid to mount root partition
Some servers assign different device labels in the installer
and firstboot which causes the boot process to break.  For example
sometimes a virtual cd might be /dev/sda during intstall, but not
be present after the reboot so now the root drive is /dev/sda.
Mounting by uuid will allow the boot to work even if the device labels
change between boots.

Signed-off-by: davidplunkett <dp7642@att.com>
Change-Id: I82025fc69285418bae48b7b32d31d3e1a0552cbb